group java classes into packages
