+++ src/main/java/qi/command/AddCommand.java
@@ -1,3 +1,14 @@
+package qi.command;
+
+import qi.qiexception.QiException;
+import qi.storage.Storage;
+import qi.task.Deadline;
+import qi.task.Event;
+import qi.task.Task;
+import qi.task.Todo;
+import qi.tasklist.TaskList;
+import qi.ui.Ui;
+
 import java.io.IOException;
 import java.time.LocalDate;
