+++ src/main/java/qi/qiexception/QiException.java
@@ -1,3 +1,5 @@
+package qi.qiexception;
+
 public class QiException extends Exception {
     public QiException(String e) {
         super(e);
