+++ src/main/java/qi/storage/Storage.java
@@ -1,3 +1,8 @@
+package qi.storage;
+
+import qi.qiexception.QiException;
+import qi.tasklist.TaskList;
+import qi.task.*;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
