+++ src/main/java/qi/task/Deadline.java
@@ -1,3 +1,5 @@
+package qi.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
