+++ src/main/java/qi/task/Event.java
@@ -1,3 +1,5 @@
+package qi.task;
+
 public class Event extends Task {
     private final String startTime;
     private final String endTime;
