+++ src/main/java/qi/task/Task.java
@@ -1,3 +1,5 @@
+package qi.task;
+
 public class Task {
     protected String task;
     protected boolean done;
