+++ src/main/java/qi/task/Todo.java
@@ -1,3 +1,5 @@
+package qi.task;
+
 public class Todo extends Task {
     public Todo(String task) {
         super(task);
